# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/meals": {
      "post": {
        "operationId": "MealsController_create",
        "summary": "Create a new meal",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMealDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Meal created successfully"
          }
        },
        "tags": [
          "meals"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "MealsController_findAll",
        "summary": "Get all meals for user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return all meals for user"
          }
        },
        "tags": [
          "meals"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/meals/date": {
      "get": {
        "operationId": "MealsController_findByDate",
        "summary": "Get meals for a specific date",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "query",
            "description": "Date in ISO format",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return meals for date"
          }
        },
        "tags": [
          "meals"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/meals/{id}": {
      "get": {
        "operationId": "MealsController_findOne",
        "summary": "Get meal by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return meal by ID"
          },
          "404": {
            "description": "Meal not found"
          }
        },
        "tags": [
          "meals"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "MealsController_update",
        "summary": "Update meal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMealDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Meal updated successfully"
          },
          "404": {
            "description": "Meal not found"
          }
        },
        "tags": [
          "meals"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "MealsController_remove",
        "summary": "Delete meal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Meal deleted successfully"
          },
          "404": {
            "description": "Meal not found"
          }
        },
        "tags": [
          "meals"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/meals/{id}/items": {
      "post": {
        "operationId": "MealsController_addMealItem",
        "summary": "Add food item to meal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMealItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Food item added to meal successfully"
          },
          "404": {
            "description": "Meal or food not found"
          }
        },
        "tags": [
          "meals"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/meals/{mealId}/items/{itemId}": {
      "delete": {
        "operationId": "MealsController_removeMealItem",
        "summary": "Remove food item from meal",
        "parameters": [
          {
            "name": "mealId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Food item removed from meal successfully"
          },
          "404": {
            "description": "Meal or food item not found"
          }
        },
        "tags": [
          "meals"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/meals/{id}/analyze-image": {
      "post": {
        "operationId": "MealsController_analyzeAndAddImage",
        "summary": "Analyze image and add meal items",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Image file to analyze"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "meals"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/goals": {
      "post": {
        "operationId": "GoalsController_create",
        "summary": "Create a new goal",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGoalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Goal created successfully"
          }
        },
        "tags": [
          "goals"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "GoalsController_findAll",
        "summary": "Get all goals for user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return all goals for user"
          }
        },
        "tags": [
          "goals"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/goals/active": {
      "get": {
        "operationId": "GoalsController_getActiveGoal",
        "summary": "Get active goal for user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return active goal"
          },
          "404": {
            "description": "No active goal found"
          }
        },
        "tags": [
          "goals"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/goals/{id}": {
      "get": {
        "operationId": "GoalsController_findOne",
        "summary": "Get goal by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return goal by ID"
          },
          "404": {
            "description": "Goal not found"
          }
        },
        "tags": [
          "goals"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "GoalsController_update",
        "summary": "Update goal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGoalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Goal updated successfully"
          },
          "404": {
            "description": "Goal not found"
          }
        },
        "tags": [
          "goals"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "GoalsController_remove",
        "summary": "Delete goal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Goal deleted successfully"
          },
          "404": {
            "description": "Goal not found"
          }
        },
        "tags": [
          "goals"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Fitness Tracker API",
    "description": "API for fitness and nutrition tracking application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateMealDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "BREAKFAST",
              "LUNCH",
              "DINNER",
              "SNACK"
            ],
            "example": "BREAKFAST"
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "example": "2023-10-15T08:00:00Z"
          }
        },
        "required": [
          "type",
          "date"
        ]
      },
      "UpdateMealDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "BREAKFAST",
              "LUNCH",
              "DINNER",
              "SNACK"
            ],
            "example": "BREAKFAST"
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "example": "2023-10-15T08:00:00Z"
          }
        }
      },
      "AddMealItemDto": {
        "type": "object",
        "properties": {
          "foodId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "quantity": {
            "type": "number",
            "example": 100
          },
          "unit": {
            "type": "string",
            "example": "g"
          }
        },
        "required": [
          "foodId",
          "quantity",
          "unit"
        ]
      },
...
      "CreateGoalDto": {
        "type": "object",
        "properties": {
          "calories": {
            "type": "number",
            "example": 2000
          },
          "protein": {
            "type": "number",
            "example": 150
          },
          "carbs": {
            "type": "number",
            "example": 200
          },
          "fat": {
            "type": "number",
            "example": 65
          }
        },
        "required": [
          "calories",
          "protein",
          "carbs",
          "fat"
        ]
      },
      "UpdateGoalDto": {
        "type": "object",
        "properties": {
          "calories": {
            "type": "number",
            "example": 2000
          },
          "protein": {
            "type": "number",
            "example": 150
          },
          "carbs": {
            "type": "number",
            "example": 200
          },
          "fat": {
            "type": "number",
            "example": 65
          }
        }
      },
...
    }
  }
}